gaze-heatmap-backend/app/requirements.txt
flask
pillow
matplotlib
numpy
scipy
flask-cors
